Utvider personoversikt (tidligere fagsakoversikt) til å vise personopplysninger og behandlingsoversikt
