+++ package.json
@@ -107,6 +107,8 @@
     "nav-frontend-spinner": "^2.0.17",
     "nav-frontend-spinner-style": "^1.0.2",
     "nav-frontend-tabell-style": "^0.2.23",
+    "nav-frontend-tabs": "^2.0.1",
+    "nav-frontend-tabs-style": "^2.0.1",
     "nav-frontend-typografi": "^3.1.1",
     "nav-frontend-typografi-style": "^1.1.1",
     "node-fetch": "^2.6.1",
