+++ src/frontend/typer/common.ts
@@ -1,0 +1,4 @@
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type PartialRecord<K extends keyof any, T> = {
+    [P in K]?: T;
+};
